modiefied files according to favourite

+++ backend/routes/favouriteRoutes.js
@@ -10,7 +10,7 @@ router.get('/:userId',authenticate, async(req,res)=>{
     try {
         const favourites=await Favourite.find({user:userId}).populate('product')
         const favouritesArray=favourites.map((fav)=>{return(fav.product)})
-        res.status(200).json(favourites);
+        res.status(200).json(favouritesArray);
     } catch (error) {
         console.error(error);
         res.status(404).json({msg:"error in getting favourites"});
@@ -19,6 +19,11 @@ router.get('/:userId',authenticate, async(req,res)=>{
 
 router.post('/',authenticate,async(req,res)=>{
     const {userId, productId}=req.body;
+
+    const isFavExist=await Favourite.findOne({user:userId,product:productId});
+    if(isFavExist){
+        return res.status(400).json({msg:`this product already exists in your favourite list`});
+    }
 
     try {
         await Favourite.create({
@@ -34,6 +39,12 @@ router.post('/',authenticate,async(req,res)=>{
 
 router.delete('/', authenticate, async(req,res)=>{
     const {userId,productId}=req.body;
+
+    const isFavExist=await Favourite.findOne({user:userId,product:productId});
+    if(!isFavExist){
+        return res.status(400).json({msg:`this product does not exists in your favourite list`});
+    }
+
     try{
         await Favourite.findOneAndDelete({user:userId,product:productId});
         res.status(200).json({msg:'deleted successfully'});
